#19 - create Triangle component
