+++ app/components/Bubble/Bubble.tsx
@@ -4,6 +4,7 @@ import styles from './style';
 import { borderRadius, colors } from '../../config/styles';
 import { widthPercentageToDP as wp } from 'react-native-responsive-screen';
 import { Svg } from 'expo';
+import Triangle,  {triangleSides} from "../Triangle/Triangle";
 
 export interface BubbleProps {
     readonly color?: string;
@@ -16,37 +17,12 @@ export interface BubbleProps {
     readonly size?: number;
 }
 
-export enum triangleSides {
-    bottomLeft = 'bottomLeft',
-    bottomRight = 'bottomRight',
-    topLeft = 'topLeft',
-    topRight = 'topRight'
-}
 
 class Bubble extends Component<BubbleProps> {
 
 
-    createTriangleShape(side: string) {
-        const { Polygon } = Svg;
-        switch(side) {
-            case triangleSides.bottomLeft:
-                return <Polygon points="0 0, 0 100, 100 0" fill="url(#grad)" />;
-            case triangleSides.bottomRight:
-                return <Polygon points="0 0, 100 0, 100 100" fill="url(#grad)" />;
-            case triangleSides.topRight:
-                return <Polygon points="0 100, 100 0, 100 100" fill="url(#grad)" />;
-            case triangleSides.topLeft:
-                return <Polygon points="0 0, 100 100, 0 100" fill="url(#grad)" />;
-            default:
-                return "";
-        }
-    }
-
     render() {
-        const { Defs, LinearGradient, Stop } = Svg;
         const { size, color, borderRadius, borderRadiusTopLeft, borderRadiusTopRight, borderRadiusBottomLeft, borderRadiusBottomRight, triangleSide, children } = this.props;
-        const triangle = triangleSide ? this.createTriangleShape(triangleSide) : null;
-
 
         return (
             <View style={
@@ -72,30 +48,7 @@ class Bubble extends Component<BubbleProps> {
                     {children}
                 </View>
                 {(triangleSide !== '' && triangleSide !== undefined) ? (
-
-                    <View
-                        style={
-                            [
-                                styles.triangle,
-                                {
-                                    left: (triangleSide === triangleSides.topLeft || triangleSide === triangleSides.bottomLeft) ? 0 : 'auto',
-                                    right: (triangleSide === triangleSides.topRight || triangleSide === triangleSides.bottomRight) ? 0 :'auto',
-                                    top: (triangleSide === triangleSides.bottomRight || triangleSide === triangleSides.bottomLeft ) ? '100%' : -size
-                                }
-                            ]
-                        }>
-                        <Svg height={size}
-                             width={size}
-                             viewBox="0 0 100 100">
-                            <Defs>
-                                <LinearGradient id="grad" x1="0" y1="0" x2="170" y2="0">
-                                    <Stop offset="1" stopColor={colors.backgroundColorDarkSecondary} stopOpacity="1"/>
-                                    <Stop offset="0" stopColor={colors.backgroundColorDark} stopOpacity="1"/>
-                                </LinearGradient>
-                            </Defs>
-                            {triangle}
-                        </Svg>
-                    </View>
+                    <Triangle size={size} triangleSide={triangleSide}/>
                 ) : null
                 }
             </View>
