+++ app/components/Bubble/style.tsx
@@ -14,14 +14,5 @@ export default StyleSheet.create({
         padding: spacing.gutter,
         width: '100%',
         overflow: 'hidden'
-    },
-    triangle: {
-        position: 'absolute',
-    },
-    triangleLeft: {
-        left: 0
-    },
-    triangleRight: {
-        right: 0
     }
 });
