+++ app/components/Triangle/style.tsx
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    triangle: {
+        position: 'absolute'
+    }
+});
